python escape text: add empty new line only when needed at the end of text
Note: always adding an extra new line after the very last line of text in order
to always close the code block of a for loop or an if is no longer needed (as
seen as an empty "...", then a ">>>" in the python interpreter). The modified
add_eol_pat takes care of this. As a consequence, you are not going to see two
empty prompts in the python interpreter (">>>" twice), when you send over
non-indented zero-level code blocks at the end of the text.