Don't pollute global namespace with ftplugin functions
As discussed here:

    https://github.com/jpalardy/vim-slime/pull/183#issuecomment-453020461

Many of these functions have names that may collide with others in the
global namespace. For example, both the Haskell and the Matlab ftplugins
define a function called Is_comment which is global, so if you open a
Haskell file then a Matlab one, the second version of the function will
clobber the other.

So, make most of the functions script-local. And the ones that are
copy-pasted between the Haskell and Matlab plug-ins, move them into a
separate autoloaded file so that they can be shared but only loaded on
first use.