Add handler for sh
If tabs are send into a bash, they trigger bash completion and that
results in the REPL not really working as expected. So we translate tabs
into spaces, which in context of shell is usually safe operation. There
might be some edge cases, but it definitely fixes more things then it
breaks.