Improve handling of multibyte characters in tmux target for vim-slime
This commit updates the tmux target implementation in vim-slime to
handle multibyte characters correctly when sending text to tmux. The
previous method of slicing the text based on byte coun could lead to
broken characters if the text contained multibyte characters. By using
`strchars` and `strcharpart`, we ensure that the text is divided based
on character count, preserving the integrity of multibyte characters.